fixed a minor bug in link creation. we need to check if it already exists. CRM-80
git-svn-id: http://svn.civicrm.org/civicrm/trunk@1326 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ php4/createlinks.sh
@@ -1,10 +1,23 @@
 #!/bin/bash
 
 for fileLink in css i js l10n packages PEAR templates
-do
-  echo creating link $fileLink ...
-  ln -s ~/svn/crm/$fileLink $fileLink
+  do
+  if test -L $fileLink   # check if link exists
+      then
+      # skip if link exists since it creates another symlink in the source directory
+      # which breaks the drupal admin calls (goes into infinite loop trying to follow circular links)
+      echo link $fileLink already exists. skipping it ...
+  else
+      echo creating link $fileLink ...
+      ln -s ~/svn/crm/$fileLink $fileLink
+  fi
 done
 
 cd modules
-ln -s civicrm.module.hide civicrm.module
+if test -L civicrm.module   # check if link exists
+    then
+    echo link civicrm.module already exists. skipping it ...
+else
+    echo creating link civicrm.module ...
+    ln -s civicrm.module.hide civicrm.module
+fi
